Struct jamsocket_server::ClientSocketConnection

pub struct ClientSocketConnection {
    pub room: Recipient<MessageFromClient>,
    pub client_id: ClientId,
    pub room_id: String,
    pub ip: String,
    pub last_seen: Instant,
    pub heartbeat_interval: Duration,
    pub heartbeat_timeout: Duration,
    pub interval_handle: Option<SpawnHandle>,
}

Represents a connection from a service to a client, which consists of a message receiver and a user ID.

Fields
room: Recipient<MessageFromClient>client_id: ClientIdroom_id: Stringip: Stringlast_seen: Instantheartbeat_interval: Durationheartbeat_timeout: Durationinterval_handle: Option<SpawnHandle>